Behat test written
