+++ server/moodle/mod/livequiz/templates/question_menu_popup.mustache
@@ -32,7 +32,7 @@
 <div class="Modal_div">
     <h3>Title</h3>
     <div id="title_textarea_alert" style="display: none;">
-        <p class="title_alert_text" id="title_alert_text_id" style="color: red">*This field is required</p>
+        <p class="title_alert_text" id="title_alert_text_id" style="color: red">*The Title is required</p>
     </div>
     <textarea
     placeholder="Enter title (ex. Geography 1)"
@@ -41,7 +41,7 @@
   ></textarea>
     <h3>Question Text</h3>
     <div id="question_textarea_alert" style="display: none;">
-        <p class="question_alert_text" id="question_alert_text_id" style="color: red" >*This field is required</p>
+        <p class="question_alert_text" id="question_alert_text_id" style="color: red" >*The Question Text is required</p>
     </div>
     <textarea
           placeholder="Enter question (ex. What is the capital of Sweden?)"
@@ -71,7 +71,7 @@
         <p class="all_answers_alert_description" id="question_alert_description" style="color: red">*Each answer must have a description</p>
     </div>
   <div class="all_answers_for_question_div"></div>
-    <p id="validText" style="color: red;display: none">Please fill out required fields</p>
+    <p id="validText" style="color: red;display: none">Please fill out all the required fields</p>
   <button class="save_button">Save Question</button>
   <button class="discard_question_button">Discard</button>
 </div>
