+++ server/moodle/mod/livequiz/tests/behat/quizcreator_livequiz.feature
@@ -148,3 +148,38 @@ Feature: View livequiz activity
     And "Stockholm" "radio" should exist
     And "Malmö" "radio" should exist
     And "Gothenburg" "radio" should exist
+
+  Scenario: Validate submission for a question
+    When I click on "livequiz_europe_quiz" "link" in the "livequiz" activity
+    And I click on "Add Question" "button"
+    And I click on "Save Question" "button"
+    Then I should see "Please fill out all the required fields"
+    And I should see "The Title is required"
+    And I should see "The Question Text is required"
+    And I should see "*You need to add at least one correct answer"
+    And I should see "*You need to have at least two answers"
+    Then I set the field "question_title_id" to "Geography 1"
+    And I click on "Save Question" "button"
+    Then I should see "Please fill out all the required fields"
+    And I should see "The Question Text is required"
+    And I should see "*You need to add at least one correct answer"
+    And I should see "*You need to have at least two answers"
+    Then I set the field "question_description_id" to "What is the Capital of Sweden?"
+    And I click on "Save Question" "button"
+    Then I should see "Please fill out all the required fields"
+    And I should see "*You need to add at least one correct answer"
+    And I should see "*You need to have at least two answers"
+    Then I click on "Add Answer" "button"
+    And I click on "Add Answer" "button"
+    And I click on "Save Question" "button"
+    Then I should see "Please fill out all the required fields"
+    And I should see "*You need to add at least one correct answer"
+    And I should see "*Each answer must have a description"
+    Then I set the field with xpath "(//input[@class='answer_input'])[1]" to "Stockholm"
+    And I set the field with xpath "(//input[@class='answer_input'])[2]" to "Norway"
+    And I click on "Save Question" "button"
+    Then I should see "Please fill out all the required fields"
+    And I should see "*You need to add at least one correct answer"
+    And I set the field with xpath "(//input[@class='answer_checkbox'])[1]" to "checked"
+    And I click on "Save Question" "button"
+    Then "Geography 1" "list_item" should exist
